requirements.txt
streamlit>=1.28.0
langchain>=0.1.0
langchain-community>=0.0.20
langchain-core>=0.1.0
langchain-chroma>=0.1.0
langchain-text-splitters>=0.0.1
gTTS>=2.4.0
SpeechRecognition>=3.10.0
transformers>=4.35.0
torch>=2.1.0
torchaudio>=2.1.0
Pillow>=10.0.0
PyPDF2>=3.0.1
PyYAML>=6.0.1
python-dotenv>=1.0.0
numpy>=1.24.0
pandas>=2.0.0
requests>=2.31.0
pydub>=0.25.1
chromadb>=0.4.0
